fix: 重大なバグ修正（4件）
1. PostgreSQL整数切り詰めバグ修正（CRITICAL）
   - src/builtins/postgres.rs:81
   - i64→i32の切り詰めによるデータ破損を修正
   - BIGINT型として正しくi64を送信

2. path/normalizeセキュリティ脆弱性修正（CRITICAL）
   - src/builtins/path.rs:162-168
   - ディレクトリトラバーサル攻撃を防止
   - 相対パスの先頭の..を保持

3. HTTP圧縮破損バグ修正（HIGH）
   - src/builtins/server/helpers.rs:32
   - src/builtins/server/middleware.rs:168
   - gzipバイナリ→文字列変換による破損を修正
   - Value::Bytesで正しく扱う

4. WebSocketランタイム問題修正（HIGH）
   - src/builtins/websocket.rs
   - グローバルTokioランタイムを使用
   - 4関数すべて（connect/send/receive/close）で安定動作

+++ src/builtins/postgres.rs
@@ -78,7 +78,7 @@ impl PostgresConnection {
         match value {
             Value::Nil => Box::new(None::<String>),
             Value::Bool(b) => Box::new(*b),
-            Value::Integer(i) => Box::new(*i as i32), // PostgreSQL INTEGER型はi32
+            Value::Integer(i) => Box::new(*i), // i64のまま送る（PostgreSQL BIGINT型）
             Value::Float(f) => Box::new(*f),
             Value::String(s) => Box::new(s.clone()),
             Value::Bytes(b) => Box::new(b.as_ref().to_vec()), // BYTEA型として送信
